+++ photutils/aperture/tests/test_aperture_common.py
@@ -1,4 +1,7 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
+"""
+This module provides base classes for aperture tests.
+"""
 
 from astropy.coordinates import SkyCoord
 from astropy.tests.helper import assert_quantity_allclose